+++ CSharp/Concentus/Celt/Tables.cs
@@ -44,7 +44,7 @@ namespace Concentus.Celt
     internal static class Tables
     {
         /* Mean energy in each band quantized in Q4 */
-        internal static readonly byte[] eMeans/*[25]*/ = {
+        internal static readonly sbyte[] eMeans/*[25]*/ = {
               103,100, 92, 85, 81,
                77, 72, 70, 78, 75,
                73, 71, 78, 74, 69,
@@ -73,62 +73,62 @@ namespace Concentus.Celt
            (inter/intra), and band number.
           The first number of each pair is the probability of 0, and the second is the
            decay rate, both in Q8 precision.*/
-        internal static readonly byte[][][] e_prob_model = {
+        internal static readonly short[][][] e_prob_model = {
            /*120 sample frames.*/
-           new byte[][]{
+           new short[][]{
               /*Inter*/
-              new byte[]{
+              new short[]{
                   72, 127,  65, 129,  66, 128,  65, 128,  64, 128,  62, 128,  64, 128,
                   64, 128,  92,  78,  92,  79,  92,  78,  90,  79, 116,  41, 115,  40,
                  114,  40, 132,  26, 132,  26, 145,  17, 161,  12, 176,  10, 177,  11
               },
               /*Intra*/
-              new byte[]{
+              new short[]{
                   24, 179,  48, 138,  54, 135,  54, 132,  53, 134,  56, 133,  55, 132,
                   55, 132,  61, 114,  70,  96,  74,  88,  75,  88,  87,  74,  89,  66,
                   91,  67, 100,  59, 108,  50, 120,  40, 122,  37,  97,  43,  78,  50
               }
            },
            /*240 sample frames.*/
-           new byte[][]{
+           new short[][]{
               /*Inter*/
-              new byte[]{
+              new short[]{
                   83,  78,  84,  81,  88,  75,  86,  74,  87,  71,  90,  73,  93,  74,
                   93,  74, 109,  40, 114,  36, 117,  34, 117,  34, 143,  17, 145,  18,
                  146,  19, 162,  12, 165,  10, 178,   7, 189,   6, 190,   8, 177,   9
               },
               /*Intra*/
-              new byte[]{
+              new short[]{
                   23, 178,  54, 115,  63, 102,  66,  98,  69,  99,  74,  89,  71,  91,
                   73,  91,  78,  89,  86,  80,  92,  66,  93,  64, 102,  59, 103,  60,
                  104,  60, 117,  52, 123,  44, 138,  35, 133,  31,  97,  38,  77,  45
               }
            },
            /*480 sample frames.*/
-           new byte[][]{
+           new short[][]{
               /*Inter*/
-              new byte[]{
+              new short[]{
                   61,  90,  93,  60, 105,  42, 107,  41, 110,  45, 116,  38, 113,  38,
                  112,  38, 124,  26, 132,  27, 136,  19, 140,  20, 155,  14, 159,  16,
                  158,  18, 170,  13, 177,  10, 187,   8, 192,   6, 175,   9, 159,  10
               },
               /*Intra*/
-              new byte[]{
+              new short[]{
                   21, 178,  59, 110,  71,  86,  75,  85,  84,  83,  91,  66,  88,  73,
                   87,  72,  92,  75,  98,  72, 105,  58, 107,  54, 115,  52, 114,  55,
                  112,  56, 129,  51, 132,  40, 150,  33, 140,  29,  98,  35,  77,  42
               }
            },
            /*960 sample frames.*/
-           new byte[][]{
+           new short[][]{
               /*Inter*/
-              new byte[]{
+              new short[]{
                   42, 121,  96,  66, 108,  43, 111,  40, 117,  44, 123,  32, 120,  36,
                  119,  33, 127,  33, 134,  34, 139,  21, 147,  23, 152,  20, 158,  25,
                  154,  26, 166,  21, 173,  16, 184,  13, 184,  10, 150,  13, 139,  15
               },
               /*Intra*/
-              new byte[]{
+              new short[]{
                   22, 178,  63, 114,  74,  82,  84,  83,  92,  82, 103,  62,  96,  72,
                   96,  67, 101,  73, 107,  72, 113,  55, 118,  52, 125,  52, 118,  52,
                  117,  55, 135,  49, 137,  39, 157,  32, 145,  29,  97,  33,  77,  40
@@ -150,7 +150,7 @@ namespace Concentus.Celt
         internal static readonly short[] tapset_icdf = { 2, 1, 0 };
 
         /* Bit allocation table in units of 1/32 bit/sample (0.1875 dB SNR) */
-        internal static readonly byte[] band_allocation = {
+        internal static readonly short[] band_allocation = {
             /*0  200 400 600 800  1k 1.2 1.4 1.6  2k 2.4 2.8 3.2  4k 4.8 5.6 6.8  8k 9.6 12k 15.6 */
               0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,
              90, 80, 75, 69, 63, 56, 49, 40, 34, 29, 20, 18, 10,  0,  0,  0,  0,  0,  0,  0,  0,
@@ -352,7 +352,7 @@ namespace Concentus.Celt
             240, 240, 305, 305, 305, 305, 343, 343, 343, 351, 351, 370, 376, 382, 387,
             };
 
-        internal static readonly byte[] cache_bits50 = {
+        internal static readonly short[] cache_bits50 = {
 40, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7,
             7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7,
             7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 40, 15, 23, 28,
@@ -381,7 +381,7 @@ namespace Concentus.Celt
             224, 5, 60, 113, 161, 206, 248, 4, 65, 122, 175, 224, 4, 67, 127,
             182, 234, };
 
-        internal static readonly byte[] cache_caps50 = {
+        internal static readonly short[] cache_caps50 = {
             224, 224, 224, 224, 224, 224, 224, 224, 160, 160, 160, 160, 185, 185, 185,
             178, 178, 168, 134, 61, 37, 224, 224, 224, 224, 224, 224, 224, 224, 240,
             240, 240, 240, 207, 207, 207, 198, 198, 183, 144, 66, 40, 160, 160, 160,
